seperated server and app js
